requirements.txt
numpy
opencv-python
fastapi
uvicorn[standard]
python-multipart
Pillow
sqlalchemy
psycopg2-binary
python-dotenv


